+++ ddl/check.lua
@@ -1072,5 +1072,8 @@ return {
     check_index_part = check_index_part,
     check_index_parts = check_index_parts,
     check_index = check_index,
-    check_field = check_field
+    check_field = check_field,
+    internal = {
+        is_callable = is_callable,
+    }
 }